+++ src/ScottPlot5/ScottPlot5 Controls/ScottPlot.WinUI/WinUIPlot.cs
@@ -118,11 +118,6 @@ public partial class WinUIPlot : UserControl, IPlotControl
         base.OnKeyUp(e);
     }
 
-    public Coordinates GetCoordinates(Pixel px, IXAxis? xAxis = null, IYAxis? yAxis = null)
-    {
-        return Plot.GetCoordinates(px, xAxis, yAxis);
-    }
-
     public float DetectDisplayScale()
     {
         // TODO: improve support for DPI scale detection